// src/utils/permissions/enums/componentEnum.ts
export enum ComponentEnum {
  //widgets
  AppLauncherWidget = 'APP_LAUNCHER_WIDGET',
  AnniversaryCelebrationWidget = 'ANNIVERSARY_CELEBRATION_WIDGET',
  AnnouncementWidget = 'ANNOUNCEMENT_WIDGET',
  BirthdayCelebrationWidget = 'BIRTHDAY_CELEBRATION_WIDGET',
  ChannelAdminsWidget = 'CHANNEL_ADMINS_WIDGET',
  ChannelLinksWidget = 'CHANNEL_LINKS_WIDGET',
  ChannelMembersWidget = 'CHANNEL_MEMBERS_WIDGET',
  ChannelRequestWidget = 'CHANNEL_REQUEST_WIDGET',
  ChannelsWidget = 'CHANNELS_WIDGET',
  EventWidget = 'EVENT_WIDGET',
  EvaluationRequestWidget = 'EVALUATION_REQUEST_WIDGET',
  NotificationSettingsWidget = 'NOTIFICATIONS_SETTINGS_WIDGET',
  ProgressTrackerWidget = 'PROGRESS_TRACKER_WIDGET',
  RecommendationWidget = 'RECOMMENDATION_WIDGET',
  TeamsWidget = 'TEAMS_WIDGET',
  UserCardWidget = 'USER_CARD_WIDGET',
}